winedbg: Avoid using 'long' types.

+++ programs/winedbg/be_arm.c
@@ -1772,7 +1772,7 @@ static BOOL be_arm_is_jump(const void* insn, ADDRESS64* jumpee)
 
 static BOOL be_arm_insert_Xpoint(HANDLE hProcess, const struct be_process_io* pio,
                                  dbg_ctx_t *ctx, enum be_xpoint_type type,
-                                 void* addr, unsigned long* val, unsigned size)
+                                 void* addr, unsigned *val, unsigned size)
 {
     SIZE_T              sz;
 
@@ -1790,7 +1790,7 @@ static BOOL be_arm_insert_Xpoint(HANDLE hProcess, const struct be_process_io* pi
 
 static BOOL be_arm_remove_Xpoint(HANDLE hProcess, const struct be_process_io* pio,
                                  dbg_ctx_t *ctx, enum be_xpoint_type type,
-                                 void* addr, unsigned long val, unsigned size)
+                                 void* addr, unsigned val, unsigned size)
 {
     SIZE_T              sz;
